+++ src/main/java/com/gerenciamento/projetoapi/service/TarefaService.java
@@ -4,13 +4,19 @@ import com.gerenciamento.projetoapi.model.Tarefa;
 import com.gerenciamento.projetoapi.repository.TarefaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TarefaService {
 
+    private final TarefaRepository tarefaRepository;
+
     @Autowired
-    private TarefaRepository tarefaRepository;
+    public TarefaService(TarefaRepository tarefaRepository) {
+        this.tarefaRepository = tarefaRepository;
+    }
 
     public Tarefa criarTarefa(Tarefa tarefa) {
         return tarefaRepository.save(tarefa);
@@ -20,16 +26,28 @@ public class TarefaService {
         return tarefaRepository.findAll();
     }
 
+    public Optional<Tarefa> buscarTarefa(Long id) {
+        return tarefaRepository.findById(id);
+    }
+
     public Tarefa atualizarTarefa(Long id, Tarefa tarefa) {
-        tarefa.setId(id);
-        return tarefaRepository.save(tarefa);
+        Optional<Tarefa> tarefaExistente = tarefaRepository.findById(id);
+        if (tarefaExistente.isPresent()) {
+            Tarefa t = tarefaExistente.get();
+            t.setDescricao(tarefa.getDescricao());
+            t.setStatus(tarefa.getStatus());
+            t.setResponsavel(tarefa.getResponsavel());
+            return tarefaRepository.save(t);
+        } else {
+            throw new IllegalArgumentException("Tarefa não encontrada.");
+        }
     }
 
     public void deletarTarefa(Long id) {
-        tarefaRepository.deleteById(id);
-    }
-
-    public Tarefa buscarTarefa(Long id) {
-        return tarefaRepository.findById(id).orElse(null);
+        if (tarefaRepository.existsById(id)) {
+            tarefaRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Tarefa não encontrada.");
+        }
     }
 }
